Reorganized to only expose the extension trait, instead of the whole submodule tree
